Adjust unit test
