+++ lib/collection/src/collection_manager/tests/mod.rs
@@ -629,52 +629,3 @@ fn test_proxy_shared_updates_same_version() {
         assert!(["small", "big"].contains(&size.as_str().unwrap()));
     }
 }
-
-#[test]
-fn test_delete_by_filter_version_bump() {
-    let dir = Builder::new().prefix("segment_dir").tempdir().unwrap();
-
-    let segment1 = build_segment_1(dir.path());
-    let segment2 = build_segment_2(dir.path());
-
-    let hw_counter = HardwareCounterCell::new();
-
-    let mut holder = SegmentHolder::default();
-
-    let _sid1 = holder.add_new(segment1);
-    let _sid2 = holder.add_new(segment2);
-
-    const DELETE_OP_NUM: u64 = 100;
-
-    assert!(
-        holder
-            .iter()
-            .all(|i| i.1.get().read().version() < DELETE_OP_NUM)
-    );
-
-    let segments = Arc::new(RwLock::new(holder));
-
-    // A filter that matches no points.
-    let filter = Filter::new_must(Condition::Field(FieldCondition::new_match(
-        "color".parse().unwrap(),
-        Match::Value(MatchValue {
-            value: ValueVariants::String("white".to_string()),
-        }),
-    )));
-
-    let deleted_count =
-        delete_points_by_filter(&segments.read(), DELETE_OP_NUM, &filter, &hw_counter).unwrap();
-    assert_eq!(deleted_count, 0); // Assert no points were deleted.
-
-    // Now we should see `DELETE_OP_NUM` as version of one of the appendable segments.
-    assert_eq!(
-        segments
-            .read()
-            .appendable_segments_ids()
-            .into_iter()
-            .filter(|i| segments.read().get(*i).unwrap().get().read().version() == DELETE_OP_NUM)
-            .count(),
-        1,
-        "must have exactly one segment with version {DELETE_OP_NUM}",
-    );
-}
